CSS attributes: centralize codegen and verify generated outputs
