+++ .cspell.json
@@ -57,6 +57,7 @@
 		"checkmark",
 		"checksummed",
 		"clineignore",
+		"codegen",
 		"Coinspect",
 		"colorlinks",
 		"Consensys",
